requirements.txt
langchain
langchain-experimental
langchain-huggingface
sentence-transformers
scikit-learn
streamlit
python-dotenv